+++ ansible_navigator/image_manager/__init__.py
@@ -1,4 +1,4 @@
 """ image manager """
 
-from .image_puller import ImagePuller
-from .image_inspector import inspect_all
+from .puller import ImagePuller
+from .inspector import inspect_all
