+++ ansible_navigator/image_manager/inspector.py
