+++ ansible_navigator/image_manager/puller.py
